v1.3

+++ settings/flimit.go
@@ -7,8 +7,8 @@ import "syscall"
 // Flimit defines the max number of watched files
 func (s *Settings) Flimit() {
 	var rLimit syscall.Rlimit
-	rLimit.Max = s.Config.Flimit
-	rLimit.Cur = s.Config.Flimit
+	rLimit.Max = uint64(s.Config.Flimit)
+	rLimit.Cur = uint64(s.Config.Flimit)
 	err := syscall.Setrlimit(syscall.RLIMIT_NOFILE, &rLimit)
 	if err != nil {
 		s.Fatal(err, "Error setting rlimit")
